+++ examples/basic/src/components/Placeholder/Placeholder.vue
@@ -64,10 +64,8 @@ export default {
 	content: '-';
 }
 </style>
-<docs>
-
+<docs lang="md">
 Another example with custom blocks inside `*.vue` files.
 
     <Placeholder type="animal"/>
-
 </docs>