+++ packages/vue-docgen-api/src/script-handlers/index.ts
@@ -1,6 +1,7 @@
 import classDisplayNameHandler from './classDisplayNameHandler'
 import classMethodHandler from './classMethodHandler'
 import classPropHandler from './classPropHandler'
+import classEventHandler from './classEventHandler'
 import componentHandler from './componentHandler'
 import displayNameHandler from './displayNameHandler'
 import eventHandler from './eventHandler'
@@ -17,6 +18,7 @@ export {
 	classDisplayNameHandler,
 	classMethodHandler,
 	classPropHandler,
+	classEventHandler,
 	componentHandler,
 	displayNameHandler,
 	eventHandler,
@@ -40,7 +42,8 @@ const defaultHandlers: Handler[] = [
 	slotHandlerLitteral,
 	classDisplayNameHandler,
 	classMethodHandler,
-	classPropHandler
+	classPropHandler,
+	classEventHandler
 ]
 
 export const preHandlers: Handler[] = [