+++ packages/vue-docgen-api/tests/components/docs-only/docs.vue
@@ -1,7 +1,5 @@
-<docs>
-
+<docs lang="md">
 ```js { "className": "checks" }
 <Button>I’m transparent!</Button>
 ```
-
 </docs>